+++ src/main/scala/uk/gov/hmrc/hmrcfrontend/views/config/StandardPhaseBanner.scala
@@ -15,11 +15,29 @@
  */
 
 package uk.gov.hmrc.hmrcfrontend.views.config
+
 import play.api.i18n.Messages
+import play.api.mvc.RequestHeader
 import play.twirl.api.HtmlFormat
 import uk.gov.hmrc.govukfrontend.views.Aliases.{HtmlContent, PhaseBanner, Tag, Text}
+import uk.gov.hmrc.hmrcfrontend.config.ContactFrontendConfig
+import uk.gov.hmrc.hmrcfrontend.views.Utils.urlEncode
 
 trait StandardPhaseBanner {
+
+  protected def contactFrontendBetaFeedbackUrl()(implicit
+    contactFrontendConfig: ContactFrontendConfig,
+    request: RequestHeader
+  ) = {
+    val queryStringParams: Seq[String] = Seq(
+      contactFrontendConfig.serviceId.map(id => "service=" + id),
+      contactFrontendConfig.referrerUrl.map(url => "referrerUrl=" + urlEncode(url))
+    ).flatten
+
+    val queryString = if (queryStringParams.isEmpty) "" else "?" + queryStringParams.mkString("&")
+    contactFrontendConfig.baseUrl.getOrElse("") + "/contact/beta-feedback" + queryString
+  }
+
   def apply(phase: String, url: String)(implicit messages: Messages): PhaseBanner =
     PhaseBanner(
       tag = Some(Tag(content = Text(phase))),
@@ -33,6 +51,13 @@ trait StandardPhaseBanner {
 
 class StandardBetaBanner extends StandardPhaseBanner {
   def apply(url: String)(implicit messages: Messages): PhaseBanner = apply(phase = "beta", url = url)
+
+  def apply()(implicit
+    contactFrontendConfig: ContactFrontendConfig,
+    request: RequestHeader,
+    messages: Messages
+  ): PhaseBanner =
+    apply(url = contactFrontendBetaFeedbackUrl())
 }
 
 class StandardAlphaBanner extends StandardPhaseBanner {