+++ src/test/scala/uk/gov/hmrc/hmrcfrontend/views/config/StandardPhaseBannerSpec.scala
@@ -19,15 +19,19 @@ package uk.gov.hmrc.hmrcfrontend.views.config
 import org.scalatest.matchers.must.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 import org.scalatestplus.play.guice.GuiceOneAppPerSuite
+import play.api.Configuration
 import play.api.i18n.{Lang, Messages}
+import play.api.mvc.RequestHeader
+import play.api.test.FakeRequest
 import uk.gov.hmrc.govukfrontend.views.Aliases.{HtmlContent, PhaseBanner, Tag, Text}
 import uk.gov.hmrc.helpers.MessagesSupport
+import uk.gov.hmrc.hmrcfrontend.config.ContactFrontendConfig
 
 class StandardPhaseBannerSpec extends AnyWordSpec with Matchers with MessagesSupport with GuiceOneAppPerSuite {
   "StandardBetaBanner" must {
     val standardBetaBanner = app.injector.instanceOf[StandardBetaBanner]
 
-    "Return the correct PhaseBanner object for the given phase and url" in {
+    "return the correct PhaseBanner object for the given phase and url" in {
       val phaseBanner = standardBetaBanner(url = "/feedback")
 
       phaseBanner mustBe PhaseBanner(
@@ -38,7 +42,7 @@ class StandardPhaseBannerSpec extends AnyWordSpec with Matchers with MessagesSup
       )
     }
 
-    "Return the correct PhaseBanner object for a different phase and url" in {
+    "return the correct PhaseBanner object for a different phase and url" in {
       val phaseBanner = standardBetaBanner(url = "/other-feedback")
 
       phaseBanner mustBe PhaseBanner(
@@ -60,7 +64,7 @@ class StandardPhaseBannerSpec extends AnyWordSpec with Matchers with MessagesSup
       )
     }
 
-    "Return the correct Welsh content" in {
+    "return the correct Welsh content" in {
       implicit val messages: Messages = messagesApi.preferred(Seq(Lang("cy")))
 
       val phaseBanner = standardBetaBanner(url = "/feedback")
@@ -72,12 +76,44 @@ class StandardPhaseBannerSpec extends AnyWordSpec with Matchers with MessagesSup
         )
       )
     }
+
+    "return formatted URL with config values for beta-feedback in contact-frontend when no URL provided" in {
+      implicit val request: RequestHeader          = FakeRequest()
+      implicit val cfConfig: ContactFrontendConfig = new ContactFrontendConfig(Configuration.empty) {
+        override def serviceId: Option[String]                    = Some("my-service")
+        override def baseUrl                                      = Some("tax.service.gov.uk")
+        override def referrerUrl(implicit request: RequestHeader) =
+          Some("/help/terms-and-conditions")
+      }
+
+      val phaseBanner: PhaseBanner = standardBetaBanner()
+      phaseBanner.content mustBe HtmlContent(
+        "This is a new service – your <a class=\"govuk-link\" href=" +
+          "\"tax.service.gov.uk/contact/beta-feedback?service=my-service&amp;referrerUrl=%2Fhelp%2Fterms-and-conditions\"" +
+          ">feedback</a> will help us to improve it."
+      )
+    }
+
+    "return formatted URL with empty config values for beta-feedback in contact-frontend when no URL provided" in {
+      implicit val request: RequestHeader          = FakeRequest()
+      implicit val cfConfig: ContactFrontendConfig = new ContactFrontendConfig(Configuration.empty) {
+        override def serviceId: Option[String]                    = None
+        override def baseUrl                                      = None
+        override def referrerUrl(implicit request: RequestHeader) = None
+      }
+
+      val phaseBanner: PhaseBanner = standardBetaBanner()
+      phaseBanner.content mustBe HtmlContent(
+        "This is a new service – your <a class=\"govuk-link\" href=" +
+          "\"/contact/beta-feedback\">feedback</a> will help us to improve it."
+      )
+    }
   }
 
   "StandardAlphaBanner" must {
     val standardAlphaBanner = app.injector.instanceOf[StandardAlphaBanner]
 
-    "Return the correct PhaseBanner object for the given url" in {
+    "return the correct PhaseBanner object for the given url" in {
       val phaseBanner = standardAlphaBanner(url = "/feedback")
 
       phaseBanner mustBe PhaseBanner(